Run.sh script changes (#16)
* change --debug flag to --variant debug

* check for bob.jar before redownloading it

* check for dmengineheadless before redownloading

* update variant argument

* headless variant

* fix typo

+++ .test/run.sh
@@ -17,15 +17,19 @@ echo "Using Defold dmengine_headless version ${SHA1}"
 DMENGINE_URL="http://d.defold.com/archive/${SHA1}/engine/${PLATFORM}/dmengine_headless"
 BOB_URL="http://d.defold.com/archive/${SHA1}/bob/bob.jar"
 
-echo "Downloading ${DMENGINE_URL}"
-curl -L -o dmengine_headless ${DMENGINE_URL}
-chmod +x dmengine_headless
+if [[ ! -f dmengine_headless ]]; then
+	echo "Downloading ${DMENGINE_URL}"
+	curl -L -o dmengine_headless ${DMENGINE_URL}
+	chmod +x dmengine_headless
+fi	
 
-echo "Downloading ${BOB_URL}"
-curl -L -o bob.jar ${BOB_URL}
+if [[ ! -f bob.jar ]]; then
+	echo "Downloading ${BOB_URL}"
+	curl -L -o bob.jar ${BOB_URL}
+fi
 
 echo "Running bob.jar"
-java -jar bob.jar --debug build
+java -jar bob.jar --variant=headless build
 
 echo "Starting dmengine_headless"
 ./dmengine_headless
